requirements.txt
streamlit==1.28.2
pandas==2.1.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
seaborn==0.13.0
schedule